Move BloomChain to separate file for readability
